Beta 2.1 predictor now functions
but poorly.
9hrs of play and near 0 improvement in the predictive network.  Rather, it seems that the only "improvements" are in built up random mutations, nothing actually has been Darwin'd out.  I plan on increasing the mutation rate and playing with the fitness calculations (either fixing them or changing).

There is still no apparent ghost markings in the "inputs" map for the predictive network even though I tried that.  I just hope that the network actually uses the history to predict and not just the current or previous frame.  I would at least expect the network to find positive results using the Bias Neuron for laying out the ground in the prediction, but no such luck, let alone actually predicting the future.